backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
slowapi>=0.1.9
pydantic>=2.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-multipart>=0.0.9
pdfplumber>=0.10.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
httpx>=0.27.0
